feat: construção do servicesAPI e implementação da requisição de receitas de comidas
