+++ package.json
@@ -48,7 +48,6 @@
     "clipboardy": "^2.1.0",
     "cypress": "^3.7.0",
     "cypress-multi-reporters": "^1.4.0",
-    "eslint": "^6.6.0",
     "eslint-config-airbnb": "^18.2.0",
     "eslint-plugin-import": "^2.22.1",
     "eslint-plugin-jsx-a11y": "^6.4.1",
